# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "自动添加设备/签名接口文档",
        "title": "苹果开发者自动添加设备/签名",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "paths": {
        "/api/deleteAppleAccount": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "删除指定苹果开发者账号",
                "parameters": [
                    {
                        "type": "string",
                        "description": "iss",
                        "name": "iss",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/deletePackage": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "删除指定IPA",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/getAllPackage": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取所有IPA",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/test": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "测试得到一个可用的苹果开发者账号",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/uploadAppleAccount": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "上传苹果开发者账号信息",
                "parameters": [
                    {
                        "type": "file",
                        "description": "p8file",
                        "name": "p8file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "iss",
                        "name": "iss",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "kid",
                        "name": "kid",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/uploadPackage": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "上传IPA",
                "parameters": [
                    {
                        "type": "file",
                        "description": "ipaFile",
                        "name": "ipaFile",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "summary",
                        "name": "summary",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}